Move cache funcs out of downloading file
