+++ src/main.rs
@@ -1,7 +1,7 @@
-mod config;
-// pub mod redis;
+pub mod cache;
+pub mod config;
 pub mod entity;
-pub mod s3;
+pub mod storage;
 pub mod styling;
 pub mod tiles;
 pub mod views;
